pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spectra-rag"
version = "0.1.0"
description = "Systematic RAG evaluation implementing 12 retrieval strategies with automated A/B testing, quality metrics, and Pareto-optimal pipeline selection."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "Spectra Contributors" },
]
keywords = [
    "rag",
    "retrieval-augmented-generation",
    "evaluation",
    "optimization",
    "nlp",
    "information-retrieval",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Indexing",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.5,<3",
    "numpy>=1.26,<3",
    "pandas>=2.1,<3",
    "sentence-transformers>=2.3,<4",
    "faiss-cpu>=1.7,<2",
    "litellm>=1.16,<2",
    "optuna>=3.4,<4",
    "scipy>=1.11,<2",
    "networkx>=3.2,<4",
    "scikit-learn>=1.3,<2",
    "rank-bm25>=0.2,<1",
    "tiktoken>=0.5,<1",
]

[project.optional-dependencies]
qdrant = ["qdrant-client>=1.7,<2"]
pgvector = ["psycopg[binary]>=3.1,<4", "pgvector>=0.2,<1"]
dashboard = ["streamlit>=1.29,<2", "plotly>=5.18,<6", "altair>=5,<6"]
dev = [
    "pytest>=7.4,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=4.1,<6",
    "ruff>=0.1,<1",
    "mypy>=1.7,<2",
]
all = ["spectra-rag[qdrant,pgvector,dashboard,dev]"]

[project.urls]
Homepage = "https://github.com/spectra-rag/spectra"
Documentation = "https://github.com/spectra-rag/spectra#readme"
Repository = "https://github.com/spectra-rag/spectra"
Issues = "https://github.com/spectra-rag/spectra/issues"

[project.scripts]
spectra = "spectra.dashboard.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/spectra"]

[tool.ruff]
target-version = "py312"
line-length = 99
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "B", "SIM", "TCH"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
